objects: separate staging
