+++ tests/func/test_data_cloud.py
@@ -231,7 +231,7 @@ def test_warn_on_outdated_stage(tmp_dir, dvc, local_remote, caplog):
 
 def test_hash_recalculation(mocker, dvc, tmp_dir, local_remote):
     tmp_dir.gen({"foo": "foo"})
-    test_get_file_hash = mocker.spy(dvc_module.oid, "get_file_hash")
+    test_get_file_hash = mocker.spy(dvc_module.objects.stage, "get_file_hash")
     ret = main(["config", "cache.type", "hardlink"])
     assert ret == 0
     ret = main(["add", "foo"])
@@ -287,7 +287,7 @@ def test_verify_hashes(
     remove("dir")
     remove(dvc.cache.local.cache_dir)
 
-    hash_spy = mocker.spy(dvc_module.oid, "get_file_hash")
+    hash_spy = mocker.spy(dvc_module.objects.stage, "get_file_hash")
 
     dvc.pull()
     assert hash_spy.call_count == 0
